Update requirements.txt
requirements.txt was looking to install/update argscomplete to its 1.10.0 version, leading to error message:

ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.                                                                                                                                                                            
censys 2.2.9 requires argcomplete<4.0.0,>=2.0.0, but you have argcomplete 1.10.0 which is incompatible.

Changed the hardcoded version to argcomplete-3.2.1 to resolve the issue and avoid the need for a manual update of the dependency after the initial install.